pyproject.toml
[project]
name = "awsurance-calculation"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "numexpr>=2.14.1",
    "pyyaml>=6.0.3",
    "sympy>=1.14.0",
]
